requirements.txt
asgiref==3.4.1
Bottleneck==1.3.4
certifi==2022.5.18.1
Django==3.2.5
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mysqlclient==2.0.3
numexpr==2.8.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pip==21.2.4
psycopg2==2.8.6
pyparsing==3.0.4
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
setuptools==61.2.0
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
wheel==0.37.1
